+++ package.json
@@ -16,7 +16,7 @@
     "axios": "^1.1.3",
     "bootstrap": "^5.2.2",
     "cookies-next": "^2.1.1",
-    "firebase": "^8.9.1",
+    "firebase": "^10.7.1",
     "html2canvas": "^1.4.1",
     "jotai": "^2.5.1",
     "moment": "^2.29.4",
